+++ coordinator/app/src/main/kotlin/lineth/coordinator/app/conflationbacktesting/ConflationBacktestingApp.kt
@@ -40,8 +40,8 @@ import lineth.coordinator.app.conflation.TracesClientFactory
 import lineth.coordinator.blockcreation.BlockCreationMonitor
 import lineth.coordinator.blockcreation.LastProvenBlockNumberProviderSync
 import lineth.coordinator.blockcreation.TargetCheckpointPauseController
+import lineth.coordinator.clients.prover.DefaultProverClientFactory
 import lineth.coordinator.clients.prover.PreRiscvProverConfig
-import lineth.coordinator.clients.prover.ProverClientFactory
 import lineth.coordinator.clients.prover.ProverConfigSwitch
 import lineth.coordinator.config.toJsonRpcRetry
 import lineth.coordinator.config.v2.CoordinatorConfig
@@ -202,8 +202,10 @@ class ConflationBacktestingApp(
       log = log,
     )
 
-  private val proverClientFactory = ProverClientFactory(
+  private val preRiscvProverClientFactory = DefaultProverClientFactory(
     vertx = vertx,
+    chainId = l2EthClient.ethChainId().get().toULong(),
+    l2MessageServiceAddress = mainCoordinatorConfig.protocol.l2.contractAddress,
     preRiscvConfig = backtestingCoordinatorConfig.preRiscvProversConfig,
     metricsFacade = metricsFacade,
   )
@@ -226,7 +228,7 @@ class ConflationBacktestingApp(
   )
 
   val proofGeneratingConflationHandlerImpl = run {
-    val executionProverClient: ExecutionProverClientV2 = proverClientFactory.preRiscvExecutionProverClient()
+    val executionProverClient: ExecutionProverClientV2 = preRiscvProverClientFactory.preRiscvExecutionProverClient()
 
     ProofGeneratingConflationHandlerImpl(
       tracesProductionCoordinator = TracesConflationCoordinatorImpl(
@@ -286,7 +288,7 @@ class ConflationBacktestingApp(
 
     val blobCompressionProofCoordinator = BlobCompressionProofCoordinator(
       vertx = vertx,
-      blobCompressionProverClient = proverClientFactory.preRiscvBlobCompressionProverClient(log = log),
+      blobCompressionProverClient = preRiscvProverClientFactory.preRiscvBlobCompressionProverClient(log = log),
       rollingBlobShnarfCalculator = RollingBlobShnarfCalculator(
         blobShnarfCalculator = GoBackedBlobShnarfCalculator(
           version = backtestingCoordinatorConfig.conflation.blobCompression.shnarfCalculatorVersion,
@@ -360,7 +362,7 @@ class ConflationBacktestingApp(
       ftxRollingInfoProvider = FtxRollingInfoProviderImpl(DisabledForcedTransactionsDao()),
     ),
     consecutiveProvenBlobsProvider = inMemoryProvenBlobsTracker,
-    proofAggregationClient = proverClientFactory.preRiscvProofAggregationProverClient(),
+    proofAggregationClient = preRiscvProverClientFactory.preRiscvProofAggregationProverClient(),
     metricsFacade = metricsFacade,
   )